ESMLogits/requirements.txt
anndata==0.8.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
biopython==1.79
biotite==0.35.0
bleach==5.0.1
brotlipy==0.7.0
build==0.9.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
clize==5.0.0
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cryptography==3.4.8
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
docutils==0.19
entrypoints==0.4
evolocity==1.0
exceptiongroup==1.1.0
executing==1.2.0
fair-esm==2.0.0
fonttools==4.38.0
h5py==3.7.0
idna==3.4
igraph==0.10.2
importlib-metadata==5.2.0
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.14.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.0.6
jupyter_core==4.12.0
keyring==23.13.1
kiwisolver==1.4.4
llvmlite==0.39.1
louvain==0.8.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
more-itertools==9.0.0
msgpack==1.0.4
natsort==8.2.0
nest-asyncio==1.5.6
networkx==2.8.8
numba==0.56.4
numpy==1.23.4
od==2.0.2
packaging==23.1
pandas==1.5.1
parso==0.8.3
patsy==0.5.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.2.2
pkginfo==1.9.2
pluggy==1.0.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pynndescent==0.5.8
pyOpenSSL==20.0.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-igraph==0.10.2
pytz==2022.6
pyzmq==19.0.2
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
scanpy==1.9.1
scikit-learn==1.0.2
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.3
session-info==1.0.0
setuptools==65.5.0
sigtools==4.0.1
six==1.16.0
stack-data==0.6.2
statsmodels==0.13.5
stdlib-list==0.8.0
texttable==1.6.4
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.3.0
umap-learn==0.5.3
urllib3==1.26.12
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.37.1
zipp==3.11.0
